Refs #292. refactored so that run_ppsd is available as mcni.run_ppsd
